Trivial cleanups.
git-svn-id: http://svn.enlightenment.org/svn/e/trunk@27184 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33

+++ e16/e/src/icccm.c
@@ -789,7 +789,8 @@ EwinSyncRequestSend(EWin * ewin)
 				 StructureNotifyMask,
 				 ECORE_X_ATOM_NET_WM_SYNC_REQUEST,
 				 Mode.events.time,
-				 count & 0xffffffff, count >> 32, 0);
+				 (long)(count & 0xffffffff),
+				 (long)(count >> 32), 0);
 
    return 1;
 }
